% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mikes_function.R
\name{mikes_function}
\alias{mikes_function}
\title{Mikes function}
\usage{
mikes_function(txt = "yep")
}
\arguments{
\item{txt}{What does mike say?}
}
\description{
This function tells you what mike says.
}
\examples{
mikes_function()
}
\keyword{output}
