% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperBlues}
\alias{paperBlues}
\title{Paper blue color palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
paperBlues
}
\description{
Paper blue color palette
}
\keyword{datasets}
