% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperDarkBlue}
\alias{paperDarkBlue}
\title{Paper dark blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
paperDarkBlue
}
\description{
Paper dark blue
}
\keyword{datasets}
