% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperDarkRed}
\alias{paperDarkRed}
\title{Paper dark red}
\format{
An object of class \code{character} of length 1.
}
\usage{
paperDarkRed
}
\description{
Paper dark red
}
\keyword{datasets}
