% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperLightBlue}
\alias{paperLightBlue}
\title{Paper light blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
paperLightBlue
}
\description{
Paper light blue
}
\keyword{datasets}
