% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperLightRed}
\alias{paperLightRed}
\title{Paper light red}
\format{
An object of class \code{character} of length 1.
}
\usage{
paperLightRed
}
\description{
Paper light red
}
\keyword{datasets}
