% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperReds}
\alias{paperReds}
\title{Paper red color palette}
\format{
An object of class \code{character} of length 9.
}
\usage{
paperReds
}
\description{
Paper red color palette
}
\keyword{datasets}
