% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\docType{data}
\name{paperRedToBlue}
\alias{paperRedToBlue}
\title{Paper red to blue radient color palette}
\format{
An object of class \code{character} of length 17.
}
\usage{
paperRedToBlue
}
\description{
Paper red to blue radient color palette
}
\keyword{datasets}
