% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{plotColors}
\alias{plotColors}
\title{A function to show colors in a plot.}
\usage{
plotColors(colors)
}
\arguments{
\item{colors}{A list of colors to plot.}
}
\description{
A function to show colors in a plot.
}
\examples{
plotColors(c("red", "white", "blue"))
}
