% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{plotTheme}
\alias{plotTheme}
\title{A customized ggplot2 theme.}
\usage{
plotTheme(theme = "basic", ...)
}
\arguments{
\item{theme}{One of 'basic', 'presentation', or 'figure'. Defaults to 'basic'.}

\item{...}{Other arguments to pass to \code{theme()}}
}
\description{
A customized ggplot2 theme.
}
\examples{
ggplot(aes(x,y)) +
    geom_point() +
    plotTheme(theme = "basic", legend.position = "none")
}
