% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{printMessage}
\alias{printMessage}
\title{Stand out message printing}
\usage{
printMessage(message = "", width = 80, fillChar = "#",
  justify = "m")
}
\arguments{
\item{message}{The message to print.}

\item{width}{Width, in characters, for the final message.}

\item{fillChar}{Character used for padding message to fill the width required.}

\item{justify}{One of "l", "m", or "r" to justify the message on the left, middle,
or right side of the output line.}
}
\description{
This function is designed to make printing flagposts and status messages easier
to see against many output messages.
}
\examples{
printMessage("hi")
}
