% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tl}
\alias{tl}
\title{Print the top left corner of a matrix or dataframe.}
\usage{
tl(obj, n = 5)
}
\arguments{
\item{obj}{A matrix or dataframe.}

\item{n}{Number of rows and columns to print.}
}
\description{
Print the first n rows and first n columns of a matrix or dataframe for a quick
look at the row/column structure of the object.
}
