% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\docType{data}
\name{cbPalette}
\alias{cbPalette}
\title{Colorblind palette}
\format{An object of class \code{character} of length 8.}
\usage{
cbPalette
}
\description{
Colorblind palette
}
\keyword{datasets}
